[getopt] Fixed our own implementation of getopt()
We have own code for getopt() that's not used now. But if we
try compiling emugen on platform without native getopt(),
e.g. Windows, we need to use ours. It had a bug where we were
not advancing the |optind| variable if option's value was in
the next |argv| item. This CL fixed it

Change-Id: I4d1f4cf226f22817fdbba28a4fbc6f34ee78ea83