[GPU Snapshot] Snapshot GL object data as it is
This patch snapshot GL object data as it is. Note that the data we
currently snapshotting are not enough for recovering hardware GPU GL
states. We need to add more data (go/oag/330989).

EglImage is not yet handled. It will affect RenderbufferData
and TextureData snapshotting (b/35390502). Experiments show that 4
TextureData are referencing EglImages during a homescreen snapshot
(nyc-emu-dev system image).

Change-Id: I2771bec93143eedb0c0cbe075a546fee408c57b4