[gl] Add a delayed color buffers cleanup.
Color buffer usage commands arrive to emulator code in
unpredictable order: sometimes we get something like this:
   (create, close, open)
It's obvious that the "open" command got out of order and was
supposed to be called on a color buffer that's alive.

We had a simple workaround to never close color buffers if
they weren't opened, but it caused memory leaks when a process
that opens a buffer dies before using it - emulator kept waiting
for an open command and never deleted abandoned color buffer.

This CL adds a separate 'delay close' list to FrameBuffer. Now
on delete we put a non-opened buffer into the delay close list,
and if nobody claimed that buffer in couple (2 so far) seconds
it gets deleted for real. On the other hand, if an open command
arives before the timeout buffer is resurrected and won't get
deleted.
It fixed the memory leak reported on the recent emulator canary
releases.

Bug: b/62441115
Change-Id: I8b2d98e18d790715af40503653dbae9f15ce9639