[GPU Snapshot] GLES3 antutu texture fix
Antutu uses glTexStorage* when GLES3 is available, which gives different
results as usng glTexImage*.

So let's add a mipmap level number to indicate whether the texture is
allocated using glTexStorage* or glTexImage*. This is because
preassigned mipmap level is the only thing that is in glTexStorage but
not in glTexImage.

Bump up snapshot version to 3.

Change-Id: I1b8ed9b6e95504156ab3685017c0655d72efc8b2