[ColorBuffer] Recreate EGL images on reformat
The ColorBuffer::reformat() should also recreate the m_eglImage. That's
because the glTexImage2D() call orphans the EGL image that was created
from it, and a brand new image is created for the texture. Hence when
what new image gets updated with new pixel content using
glTexSubImage2D(), it's not observed by the consumers of the EGL image.

bug: 64724103
Change-Id: I09cf3558e9865be211d8509691d36a1071eb482f