[GPU Snapshot] Minimum GLES1 support for API 16
This patch adds minimum GLES1 snapshot framework for API 16. After this
patch, it can start contexts without crashing. But it does not restore
most of the interesting information in contexts, so it does not render
anything after loading from a snapshot.

This CL also extract an initGlobal from context init to initialize
dispatcher explicitly. This is because GLES1 context initialization can
be skipped if the contexts are loaded from a snapshot. If we initialize
dispatcher in context initialization workflow, it could be skipped when
loading a snaphsot.

Change-Id: I59acd17032bd397f0dcecbb49ae23493d2f874d8