[gl] Consistent emulator window color on startup
bug: 69639788

With a separate posting thread, if there's nothing to swapBuffers,
the OpenGL window can show up in a variety of ways on different
platforms and backends: black, white, undefined data, a snippet
of what used to be there, etc.

This CL makes what is shown consistent by doing a glClear /
eglSwapBuffers in that case.

Change-Id: I6edac822ce1560d8867bdb298ff0e67585138261