Virtual Sensors and WASD/Mouselook integration
Update the accelerometer widget to use the current position/rotation
instead of the target, so that it animates based on the current value of
the sensors.  This integrates better with how WASD sets the current
velocity instead of target position, because the physical model would
previously only update the accelerometer widget when the target changed.

To keep the accelerometer widget fluid, the render rate has been changed
to 30 FPS.

Additionally, simplify the interaction between the physical model,
accelerometer control, and virtual sensors sliders: Previously, the
sensors page would attempt to detect when one changed and update the
other two.  Now, the physical model is considered the source of truth,
and any changes are propagated through it.

Change-Id: I2ea3ec113327d705f5a654ffec57ae3e53dfc36a